Configure request promotion patterns separate from build key
Refactor BuildKeyPromotionPatterns to make it more generic so it can be
used for more than just "build key" ordering.

Create a new static / env var for configuring the patterns that will be
used reorder requests, so this can be configured independently of the
build key ordering configuration.

Start conservatively by only promoting "*platform*" requests, and revert
the build key configuration back to its original value.
